Fix panda config, export panda preset as part of theming
