+++ .changeset/old-adults-admire.md
@@ -1,0 +1,5 @@
+---
+"@khanacademy/wonder-blocks-theming": minor
+---
+
+Add Panda preset
