+++ panda.config.ts
@@ -1,46 +1,5 @@
-import {defineConfig, defineKeyframes, defineTokens} from "@pandacss/dev";
-import {color, spacing} from "@khanacademy/wonder-blocks-tokens";
-import buttonDefaultTheme from "./packages/wonder-blocks-button/src/themes/default";
-import buttonKhanmigoTheme from "./packages/wonder-blocks-button/src/themes/khanmigo";
-
-const wbTokenToPandaToken = (token: Record<string, string | number>) =>
-    Object.entries(token)
-        .map(([key, value]) => {
-            if (typeof value === "number") {
-                value = `${value}px`;
-            }
-            return {
-                [key]: {
-                    value,
-                },
-            };
-        })
-        .reduce((acc, curr) => ({...acc, ...curr}), {}) as {
-        [x: string]: {
-            value: string;
-        };
-    };
-
-// Map WB tokens to Panda CSS tokens
-const tokens = defineTokens({
-    colors: wbTokenToPandaToken(color),
-    spacing: wbTokenToPandaToken(spacing),
-});
-
-// Supported keyframes
-const keyframes = defineKeyframes({
-    spin: {
-        "0%": {
-            transform: "rotate(0deg)",
-        },
-        "50%": {
-            transform: "rotate(180deg)",
-        },
-        "100%": {
-            transform: "rotate(360deg)",
-        },
-    },
-});
+import {defineConfig} from "@pandacss/dev";
+import {pandaPreset} from "@khanacademy/wonder-blocks-theming";
 
 export default defineConfig({
     // Whether to use css reset
@@ -52,43 +11,18 @@ export default defineConfig({
         "./__docs__/**/*.{ts,tsx,mdx}",
         "./.storybook/**/*.{ts,tsx}",
     ],
-    jsxFramework: "react",
+    // ISSUE(juan): Enabling this causes the keyframes tokens to stop working.
+    // jsxFramework: "react",
 
     // Files to exclude
     exclude: [],
 
     // Disable default tokens
-    presets: [],
-
-    // Useful for theme customization
-    theme: {
-        tokens: tokens,
-        keyframes,
-    },
-    themes: {
-        /**
-         * The themes available to the Button component.
-         */
-        // NOTE: This generates css variables at build time for each theme, so
-        // we don't have to dynamically load the theme at runtime.
-        buttonDefault: buttonDefaultTheme,
-        buttonKhanmigo: buttonKhanmigoTheme,
-    },
+    // Use WB presets instead
+    presets: [pandaPreset],
 
     // The output directory for your css system
     outdir: "styled-system",
-
-    staticCss: {
-        css: [
-            {
-                properties: {
-                    background: ["*"],
-                },
-            },
-        ],
-        themes: ["buttonDefault", "buttonKhanmigo"],
-    },
-
     importMap: "@/styled-system",
     strictTokens: false,
     // optimizations
